feat	: :zap: issue fetch나 get 시 user 정보 누적
+ user 정보를 모아두었다가 Assignee 선택, filter 시 사용
