+++ iOS/IssueTracker/IssueTracker/04.Models/Issue.swift
@@ -247,72 +247,3 @@ extension Issue {
         self.comments = comments
     }
 }
-
-// TODO: User DataModel
-struct User {
-    let name: String
-    let imageUrl: String?
-    
-    init(id: Int) {
-        name = String(id)
-        imageUrl = nil
-    }
-    
-    init(name: String, imageUrl: String? = nil) {
-        self.name = name
-        self.imageUrl = imageUrl
-    }
-    
-    init?(json: [String: Any]) {
-        guard let name = json["userName"] as? String else { return nil }
-        self.name = name
-        self.imageUrl = json["imageURL"] as? String
-    }
-}
-
-// TODO: Comment DataModel? <- 그냥 이슈가 들고있는게 좋나?
-struct Comment {
-    let content: String
-    let createAt: String
-    let author: User
-    
-    init(content: String, user: User) {
-        self.content = content
-        self.createAt = ""
-        self.author = user
-    }
-    
-    init?(json: [String: Any]?) {
-        guard let json = json,
-            let content = json["content"] as? String,
-            let createAt = json["createAt"] as? String,
-            let userObject = json["user"] as? [String: Any],
-            let user = User(json: userObject)
-            else { return nil }
-        self.content = content
-        self.createAt = createAt
-        self.author = user
-    }
-    
-}
-
-extension Comment {
-    
-    // For AddResponse
-    static func addResponse(json: [String: Any]?) -> Comment? {
-        guard let json = json,
-            let content = json["content"] as? String,
-            let createAt = json["createAt"] as? String,
-            let authorId = json["userId"] as? Int
-            else { return nil }
-        
-        return Comment(content: content, authorId: authorId, createAt: createAt)
-    }
-    
-    init(content: String, authorId: Int, createAt: String) {
-        self.content = content
-        self.createAt = createAt
-        self.author = User(id: authorId)
-    }
-    
-}
